backend/pyproject.toml
[project]
name = "mpu-platform-backend"
version = "0.1.0"
description = "MPU platform API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "alembic>=1.13.2",
  "pydantic-settings>=2.4.0",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt>=2.9.0",
  "slowapi>=0.1.9",
  "structlog>=24.4.0",
  "openai>=1.0.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "pgvector>=0.3.5",
  "stripe>=10.8.0",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.23.8",
  "faker>=28.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
